requirements.txt
kafka-python==2.3.1
kafka-python-ng==2.2.3
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
